backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beanie==1.23.3
motor==3.3.2
pydantic==2.5.0
pywebpush==1.14.0
python-dotenv==1.0.0
numpy==1.26.2
cryptography==41.0.7

